requirements.txt
streamlit 
groq
replicate
gTTS
python-dotenv